Classes/DeclarationCompatibility: code simplification [1] / deprecate `public` `$checkClassesGroups` property
Okay, so the class contained a `public` `$checkClassesGroups` property with as its value a multi-level array.

Sniff properties which are `public` can be set/overruled from a custom PHPCS ruleset to allow for customizing sniff behaviour, but multi-level property declarations are not supported for setting a property from a custom PHPCS ruleset, at least not without adding extensive special handling for these within the sniff code.

Independently of that, looking at the `$checkClassesGroups` property and the intentions for the sniff, it does not appear that this property was ever **_intended_** to be overrulable.

Now, I don't expect anyone to have overruled the property, but removing a `public` property should be considered a BC-break even so.
Even more so, as since PHP 8.2 deprecated dynamic properties, PHPCS will show a warning to users if their ruleset attempts to set a property which wasn't explicitly declared.

So, what to do ?

Well, with this commit, I propose the following:
* Deprecate the `$checkClassesGroups` property and render it unused by the sniff.
    The deprecation should be added to the action list in ticket 849 of deprecated code to remove in VIPCS 4.0.0.
* Introduce a new `private` `$extendedClassToSignatures` property with a simplified array format for the information the original property contained.
* Start using the new property in the sniff code, which allows for some code simplification.

Related to 234