Classes/DeclarationCompatibility: code simplification [2] / deprecate `public` `$checkClasses` property
Similar to the previous commit, the class contained a `public` `$checkClasses` property with as its value a complex multi-level array.

Sniff properties which are `public` can be set/overruled from a custom PHPCS ruleset to allow for customizing sniff behaviour, but multi-level property declarations are not supported for setting a property from a custom PHPCS ruleset, at least not without adding extensive special handling for these within the sniff code.

Independently of that, looking at the `$checkClasses` property and the intentions for the sniff, it does not appear that this property was ever **_intended_** to be overrulable.

Now, I don't expect anyone to have overruled the property, but removing a `public` property should be considered a BC-break even so.
Even more so, as since PHP 8.2 deprecated dynamic properties, PHPCS will show a warning to users if their ruleset attempts to set a property which wasn't explicitly declared.

Aside from all this, the array format of the property was also inconsistent - sometimes a parameter name was an array key with an array value, sometimes, the parameter name was the array value.

Having to take both situations into account makes the code more complicated than it needs to be and significantly raises the cognitive load for anyone working on this sniff.

So, what to do ?

Well, with this commit, I propose the following:
* Deprecate the `$checkClasses` property and render it unused by the sniff.
    The deprecation should be added to the action list in ticket 849 of deprecated code to remove in VIPCS 4.0.0.
* Introduce a new `private` `$methodSignatures` property with a more consistent array format for the information the original property contained.
    Parameter names are now always array keys and the value will always be an array, albeit sometimes an empty array.
* Start using the new property in the sniff code, which allows for some code simplification.

Related to 234