Classes/DeclarationCompatibility: bug fix - reference mismatch not reported
If the parent class specifies a parameter by reference and the child class did not, the sniff would report an incompatible signature.

However, if the child class specified a parameter by reference, but the parent class did not, the sniff would fail to report the incompatible signature, while this is just as much a signature mismatch in PHP. See: https://3v4l.org/hLfbR#veol

Fixed now.

Includes tests.