Classes/DeclarationCompatibility: bug fix - properly handle PHP 5.6+ variadic parameters
As things were, the sniff would verify that if a parent method would expect a parameter to be variadic, the child method would also declare the parameter as variadic.

This check is insufficient and leads to both false positives and false negatives.

Based on testing this extensively, I've boiled the rules down to this:
1. Variadic parameters must always be the last parameter in a signature.
2. If a parent method has the last parameter as optional by declaring a default value, a child method which makes that last parameter variadic and removes the default value, still complies with the signature requirements as variadic parameters are optional by nature.
3. If a parent method has the last parameter as variadic, the child method must also have the last parameter as variadic, but the child method _may_ insert additional optional parameters before the variadic parameter.

This commit adjusts the sniff code to apply all three of these rules properly.

Includes tests.
Tests for the third rule are, in part, included in the tests added in the previous commit in the `*OkayExtraOptionalParams` classes.