Classes/DeclarationCompatibility: remove redundant property
Remove a property which keeps track of the current "state". These type of properties are error prone and in this case not necessary anyway as the value of the property is only used in one place and could just have easily been retrieved where it is actually needed.

Include minor reformatting of a few long function calls for readability.