Classes/DeclarationCompatibility: bug fix - false positives for nested functions / refactor
As things were, the `DeclarationCompatibilitySniff` class extended the PHPCS native `AbstractScopeSniff`.

The `AbstractScopeSniff` has a fundamental problem, in that it will call the child sniff for every target token in every target scope.

To illustrate, let's have a look at the tests I've added:
```php
class WidgetOkayNestedFunctionIsNotMethod extends WP_Widget {
	public function hasNested() {
		// Functions can be declared nested within other functions. That doesn't make them methods.
		function is_preview($extra_param) {}
	}
}
```
As both the `hasNested()` function as well as the `is_preview()` function are nested somewhere in a `T_CLASS` scope, the `DeclarationCompatibilitySniff` will be called for both `T_FUNCTION` tokens and give both the `T_CLASS` token for the `WidgetOkayNestedFunctionIsNotMethod` class as the "current scope".

Along the same lines, for the second code sample:
```php
class WidgetOkayNestedClassDoesNotExtend extends WP_Widget {
	function hasNestedClass() {
		return new class {
			public function display_callback() {}
		};
	}
}
```
The `DeclarationCompatibilitySniff::processTokenWithinScope()` method will be called for both the `hasNestedClass()` as well as the `display_callback()` functions and both will have the `T_CLASS` token for the `WidgetOkayNestedClassDoesNotExtend` class as the "current scope", no matter that the `display_callback()` method is in actual fact a method on the anonymous class and not on the `WidgetOkayNestedClassDoesNotExtend` class.

As can be expected, this leads to false positives as both the (global) `is_preview()` function from the first code sample as well as the `display_callback()` method in the anonymous class from the second code sample would be analyzed as if they were methods on the top-level class.

Now, this can be fixed by adding a check at the start of the `processTokenWithinScope()` function to make sure that the "current scope" is the "deepest" valid scope, however, there is a better solution, which should also improve the performance of the sniff.

As of PHPCSUtils 1.1.0, PHPCSUtils offers a highly optimized `ObjectDeclarations::getDeclaredMethods()` method, which will return an array with the names of all found methods and the stack pointer to the `T_FUNCTION` keyword for each.

So now, instead of the `AbstractScopeSniff` being called for every single function and needing to do processing before deciding whether to bow out or to pass off to child classes, we can sniff for the - much rarer - `T_CLASS` token instead, then check if the class extends one of the classes we are looking for and if so, retrieve the list of methods on that class via the new PHPCSUtils method and just loop through them.

As the results for the PHPCSUtils method are cached, this also means that if _multiple_ sniffs would use the `ObjectDeclarations::getDeclared*()` methods, the result for subsequent calls will be returned nearly instantaneous, giving a PHPCS run an even bigger performance boost.