Performance/FetchingRemoteData: remove redundant method
This method was introduced in PR 855 as a temporary stop-gap to allow for supporting PHP 8.1+ first class callables.

As of WPCS 3.2.0, first class callables are no longer passed to the `process_no_parameters()` method, but to the new dedicated `process_first_class_callable()` method, so this method is no longer needed.