Security/EscapingVoidReturnFunctions: extend AbstractFunctionParameterSniff
As things were, the determination of whether or not a `T_STRING` is a call to any of the global WP `esc_*()` or `wp_kses*()`  functions, was severely flawed.

By switching the sniff over to be based on the WordPressCS `AbstractFunctionParameterSniff` class, this flaw is mitigated.

Take note of the `$target_functions` using wildcards to match all function calls starting with the prefixes indicates (as it was before).

Includes adding a slew of additional tests, some of which (line 8 - 15, line 52, 60, 63) are specific to the flaw being addressed in this commit.

Additionally, the tests have been made more comprehensive and varied by:
* Testing against false positives for calls to methods or namespaced function calls (= the issue being addressed in this PR).
* Ensure function import `use` statements are not flagged. We're not interested in those.
* Safeguarding that function calls using PHP 5.6+ argument unpacking are not flagged.
* Safeguarding handling of PHP 8.0+ attribute class using the same name as a target function.
* Safeguarding that the function is not flagged when used as a PHP 8.1+ first class callable.
* Adding tests with more variations:
    - Variation in the escaping functions being called.
    - Variation in the printing functions being called.
    - Non-lowercase function call(s), for both the escaping functions as well as the printing functions.
        This was previously handled correctly for the printing functions (via WPCS), but not for the escaping functions.
    - Fully qualified function calls for the escaping functions.
    - Use PHP 7.3+ trailing comma's in a few function calls.
    - Multi-line function call.
    - Comments in unexpected places.