Security/EscapingVoidReturnFunctions: bug fix - false positive for printing, not functions
As things were, the sniff would presume that if the first non-empty token within the escaping function call was a `T_STRING`, it would be a function call.

As the `T_STRING` token is used for quite a few things, this would lead to false positives.

Fixed now by making sure there is an open parenthesis after the `T_STRING`.

Includes test.

Ref:
* https://developer.wordpress.org/reference/functions/tag_escape/