Functions/RestrictedFunctions::is_targetted_token(): improve the method
The `RestrictedFunctions::is_targetted_token()` method was basically a duplicate of the upstream/parent method with some additional code to handle one specific situation (method calls on a specific variable).

However, the parent method has been updated significantly since the code was copied and this sniff wasn't benefitting from that.

This commit rewrites the `RestrictedFunctions::is_targetted_token()` method to defer to the parent method in all cases, except for the one specific situation we want to account for.

Additionally, it stabilizes and improves the token walking done for that specific situation, as well as takes the PHP 8.0+ nullsafe object operator into account.

Includes a number of tests, some to cover the improvements inherited from the parent method, some to cover the improved code in the overloaded method.

Includes documenting how PHP 8.1+ first class callables are handled by the sniff.