Performance/CacheValueOverride: bug fix - false positives for variables in different scopes
As things were, the searching for the "next use of the variable to which the cache was assigned" did not limit itself to the current variable scope.

In effect, that meant that "function local" variables could be mistaken for global variables and visa versa, as demonstrated by the new tests.

Fixed now.

Includes tests.